requirements.txt
discord.py
google-generativeai
Pillow
pytesseract
textwrap